requirements.txt
feedparser
openai
requests
python-dotenv
fastapi
uvicorn
python-multipart
psycopg2-binary
beautifulsoup4